+++ csvviz/cmds/line.py
@@ -8,14 +8,13 @@ $ csvviz line examples/tonk.csv --json -x date:T
 
 import altair as alt
 import click
-from csvviz.cli_utils import clout, clerr, clexit
-from csvviz.cli_utils import standard_options_decor
 from csvviz.exceptions import *
-from csvviz.vizkit import Vizkit, get_channel_name
+from csvviz.vizkit import Vizkit
 
 
 class Linekit(Vizkit):
     viz_type = "line"
+    viz_info = f"""A line chart TK"""
 
     def prepare_channels(self):
         channels = self._channels_init(self.channel_kwargs)