+++ csvviz/exceptions.py
@@ -5,3 +5,8 @@ class InvalidColorScheme(ValueError):
 class InvalidDataReference(ValueError):
     def __init__(self, msg):
         super().__init__(f"InvalidDataReference: {msg}")
+
+
+class MissingDataReference(ValueError):
+    def __init__(self, msg):
+        super().__init__(f"MissingDataReference: {msg}")